Added noncritical_index option to acts_as_solr.
When acts_as_solr is called with the :noncritical_index option set to
true the #save method for instances of that model class will *not*
fail if something goes wrong adding the record to Solr.  This is
convenient when the solr index is a noncritical part of the system.
When this option is set to true a message will be written to the log
and the save process will be allowed to continue as if the changes had
been added to solr.