Stopped orphaned entries in solr from raising exceptions.
Changed the way solr results are converted into AR objects such that
missing records in the DB do not raise an exception.  It might be
better to just remove the orphaned entry from the index as soon as we
notice but that raises some potential performance issues I don't want
to think about right now.